# expression: implement vectorized evaluation  for `builtinArithmeticPlusIntSig `