@XiaTianliang 2019-10-09
Can I get some clues from ci why the test failed, I couldn't figure it out from details。