@SunRunAway 2019-10-09
Hi, @XiaTianliang 
Thanks for your PR, it looks awesome, I've left some comments here, PTAL.
Besides, it's more appreciate adding a benchmark result in your PR description, and remove the items that are not applicable in the check list.