@XiaTianliang 2019-10-11
Finally, all test passed, I should have run **make test** before the pull request.