@XiaTianliang 2019-10-11
should i rebase master if current branch is out-of-date with base branch?