@SunRunAway 2019-10-11
> should i rebase master if current branch is out-of-date with base branch?

You do not need to, we use squash and merge to do this.